requirements.txt
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4
pydantic==1.10.13
faiss-cpu
sentence-transformers
streamlit
python-dotenv
pypdf



